final 2

+++ resources/views/shopping.blade.php
@@ -78,18 +78,20 @@ new post
                 <div class="col-2 pt-5">
                     <div class="card h-100">
 
-                        @if (auth()->user()->isAdmin)
+                        @if (isset(auth()->user()->isAdmin))
+                            @if (auth()->user()->isAdmin)
 
-                        <div class="d-flex gap-2 mx-auto " >
+                            <div class="d-flex gap-2 mx-auto " >
 
-                            {!! Form::open(['route' => ['m-book.edit' , $book->id ]  , "class" => "btn  p-0 btn-primary" , "method" => "get"]) !!}
-                            <td><button  class="btn p-1 btn-primary">Update</button></td>
-                            {!! Form::close() !!}
+                                {!! Form::open(['route' => ['m-book.edit' , $book->id ]  , "class" => "btn  p-0 btn-primary" , "method" => "get"]) !!}
+                                <td><button  class="btn p-1 btn-primary">Update</button></td>
+                                {!! Form::close() !!}
 
-                            {!! Form::open(['route' => ['m-book.destroy' , $book->id ]  , "class" => "btn  p-0 btn-danger" , "method" => "delete"]) !!}
-                            <td><button  class="btn p-1 btn-danger">Delete</button></td>
-                            {!! Form::close() !!}
-                        </div>
+                                {!! Form::open(['route' => ['m-book.destroy' , $book->id ]  , "class" => "btn  p-0 btn-danger" , "method" => "delete"]) !!}
+                                <td><button  class="btn p-1 btn-danger">Delete</button></td>
+                                {!! Form::close() !!}
+                            </div>
+                            @endif
                         @endif
 
                         <form action="{{route('m-book.show' , $book->id )}}" method="get">
@@ -98,9 +100,6 @@ new post
                                 <img src="{{ $book->image  }}" class="card-img-top" alt="Pictuere Error">
                             </button>
                           </form>
-
-
-
 
                         <div class="card-body">
                             <h5 class="card-title">{{ $book->title  }}</h5>
@@ -113,21 +112,46 @@ new post
                             <div class="d-flex gap-2 col-6 mx-auto">
                             <!-- buy -->
                             <div class="d-grid gap-2 col-6 mx-auto">
-                              <form action="{{route('cart.store')}}" method="post">
+
+
+                            @auth
+                            <form action="{{route('cart.store')}}" method="post">
                                 @csrf
                                 <input type="hidden" name="book_id" value="{{ $book->id}}">
                                 <input type="submit" value="buy" class="btn btn-warning ">
-                              </form>
+                            </form>
+                            @endauth
+
+
+                            <form action="{{route('login')}}" method="get">
+                                @csrf
+
+                                <input type="submit" value="buy" class="btn btn-warning ">
+                            </form>
+
+
                             </div>
                             <br>
                             <!-- Fav -->
                             <div class="d-grid gap-2 col-6 mx-auto">
+                                @auth
 
                             {!! Form::open(['route' => ['favourites.store' , $book->id ]  , "method" => "POST"]) !!}
 
                               {{-- <form action="{{route('favourites.create')}}" method="post"> --}}
                                 {{-- @csrf --}}
                                 <input type="hidden" name="book_id" value="{{ $book->id}}">
+                                <button type="submit" class="btn btn-outline-danger ">
+                                <svg xmlns="http://www.w3.org/2000/svg" width="25" height="25" fill="currentColor" class="bi bi-heart" viewBox="0 0 16 16"><path d="m8 2.748-.717-.737C5.6.281 2.514.878 1.4 3.053c-.523 1.023-.641 2.5.314 4.385.92 1.815 2.834 3.989 6.286 6.357 3.452-2.368 5.365-4.542 6.286-6.357.955-1.886.838-3.362.314-4.385C13.486.878 10.4.28 8.717 2.01L8 2.748zM8 15C-7.333 4.868 3.279-3.04 7.824 1.143c.06.055.119.112.176.171a3.12 3.12 0 0 1 .176-.17C12.72-3.042 23.333 4.867 8 15z"></path></svg>
+                                </button>
+                              {{-- </form> --}}
+                            {!! Form::close() !!}
+                            @endauth
+                            {!! Form::open(['route' => ['login']  , "method" => "get"]) !!}
+
+                              {{-- <form action="{{route('favourites.create')}}" method="post"> --}}
+                                {{-- @csrf --}}
+                                {{-- <input type="hidden" name="book_id" value="{{ $book->id}}"> --}}
                                 <button type="submit" class="btn btn-outline-danger ">
                                 <svg xmlns="http://www.w3.org/2000/svg" width="25" height="25" fill="currentColor" class="bi bi-heart" viewBox="0 0 16 16"><path d="m8 2.748-.717-.737C5.6.281 2.514.878 1.4 3.053c-.523 1.023-.641 2.5.314 4.385.92 1.815 2.834 3.989 6.286 6.357 3.452-2.368 5.365-4.542 6.286-6.357.955-1.886.838-3.362.314-4.385C13.486.878 10.4.28 8.717 2.01L8 2.748zM8 15C-7.333 4.868 3.279-3.04 7.824 1.143c.06.055.119.112.176.171a3.12 3.12 0 0 1 .176-.17C12.72-3.042 23.333 4.867 8 15z"></path></svg>
                                 </button>
@@ -145,10 +169,12 @@ new post
                 <div class="col-3 pt-5"> there is no such a data like u searched for . </div>
                 @endif
 
-                @if (auth()->user()->isAdmin)
-                {!! Form::open(['route' => ['m-book.create']  , "class" => "btn  btn-danger" , "method" => "get"]) !!}
-                <td><button  class="btn  btn-danger">ADD BOOK</button></td>
-                {!! Form::close() !!}
+                @if (isset(auth()->user()->isAdmin))
+                    @if (auth()->user()->isAdmin)
+                    {!! Form::open(['route' => ['m-book.create']  , "class" => "btn  btn-danger" , "method" => "get"]) !!}
+                    <td><button  class="btn  btn-danger">ADD BOOK</button></td>
+                    {!! Form::close() !!}
+                    @endif
                 @endif
